requirements.txt
# Full requirements if you also want local analysis and a simple API dashboard
# Core SDK
hyperliquid-python-sdk>=0.18.0
eth-account>=0.10
requests
python-dotenv>=1.0.0
websockets>=12.0

# Optional analytics and API server
# pandas>=2.2
# aiosqlite>=0.20
# fastapi>=0.115
# uvicorn[standard]>=0.30
